* To prevent common misconfiguration, added assertOnOffDefined() * SampleTestJob pipelines were tuned so they will always fail when light is off or settling is faulty. * Moved Machine State actuation from AbstractActuator to ReferenceActuator as the former was getting a bit fat and not so abstract.
